refactor and rename guildMemberAdd.js
Your code was not following the format of your handler. Please look at the differences before pushing this into your master branch.
